lua::IState::collectGarbage() and several minor fixes.
- Added an assert() to all __Bootstrap.cpp files to catch an unbalanced number of moduleRetain()/moduleRelease() calls.
- Added method lua::IState::collectGarbage().
- Fixed a few warnings on MSVC.

+++ src/core/Member.h
@@ -29,9 +29,9 @@ public:
 	void setOwner( ICompositeType* owner, size_t index )
 	{
 		_owner = owner;
-		
+
 		assert( index < MAX_UINT32 );
-		_index = static_cast<uint32>( index );
+		_index = static_cast<uint16>( index );
 	}
 
 	// co::IMember methods:
